requirements.txt
click==8.1.3
importlib-metadata==6.6.0
Pillow==9.5.0
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
PyQt6-sip==13.5.1
pyqt6-tools==6.4.2.3.3
python-dotenv==0.21.1
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
typing_extensions==4.5.0
zipp==3.15.0
